Structure refactored.
