+++ src/main/scala-2.12/com/bwsw/cloudstack/pulse/validators/complex/TimeScaleValidator.scala
@@ -1,6 +1,7 @@
-package com.bwsw.cloudstack.pulse.validators
+package com.bwsw.cloudstack.pulse.validators.complex
 
 import com.bwsw.cloudstack.pulse.config.PulseConfig
+import com.bwsw.cloudstack.pulse.validators.Validator
 import com.bwsw.cloudstack.pulse.validators.primitive.TimeFrameValidator
 
 /**
